ROOT doesn't support the long double type so the event time in the trees were nonsense when being read back.

+++ include/EventBuilder.hh
@@ -3,6 +3,7 @@
 
 #include <fstream>
 #include <iostream>
+#include <iomanip>
 #include <sstream>
 #include <string>
 #include <vector>
@@ -178,33 +179,33 @@ private:
 	bool event_open; ///< Flag for deciding whether an event is currently being recorded or not
 
 	// Time variables
-	long double		time_diff;	///< Time difference between first hit in event and current hit
-	long double		time_prev;	///< Holds time of previous event
-	long double		time_min;	///< The minimum time in an event containing hits
-	long double		time_max;	///< The maximum time in an event containing hits
-	long double		time_first;	///< Time of the first caen/asic/info event in a file
-	long double		ebis_prev;	///< Holds time of previous ebis pulse
-	long double		t1_prev;	///< Holds time of previous T1 pulse
-	long double		sc_prev;	///< Holds time of previous SuperCycle pulse
-	long double		laser_prev;	///< Holds time of previous Laser status pulse
-	long double		caen_time;	///< Time from the caen DAQ
-	long double		caen_prev;	///< Holds previous time from the CAEN DAQ
-	std::vector<long double> fpga_time; 			///< FPGA time on a given module of the array
-	std::vector<long double> fpga_prev;				///< Previous FPGA time on a given module of the array
-	std::vector<long double> asic_time;				///< ASIC time on a given module of the array
-	std::vector<long double> asic_prev;				///< Previous ASIC time on a given module of the array
-	std::vector<long double> pause_time;			///< The pause time on a given module of the array
-	std::vector<long double> resume_time;			///< The resume time on a given module of the array
-	std::vector<long double> asic_dead_time;		///< ASIC dead time for a given module of the array
-	std::vector<long double> asic_time_start;		///< Holds the time of the first hit on each asic in the input time-sorted tree (index denotes asic module)
-	std::vector<long double> asic_time_stop;		///< Holds the time of the last hit on each asic in the input time-sorted tree (index denotes asic module)
-	std::vector<long double> caen_time_start;		///< Holds the time of the first hit on each caen in the input time-sorted tree (index denotes caen module)
-	std::vector<long double> caen_time_stop;		///< Holds the time of the last hit on each asic in the input time-sorted tree (index denotes caen module)
+	double		time_diff;	///< Time difference between first hit in event and current hit
+	double		time_prev;	///< Holds time of previous event
+	double		time_min;	///< The minimum time in an event containing hits
+	double		time_max;	///< The maximum time in an event containing hits
+	double		time_first;	///< Time of the first caen/asic/info event in a file
+	double		ebis_prev;	///< Holds time of previous ebis pulse
+	double		t1_prev;	///< Holds time of previous T1 pulse
+	double		sc_prev;	///< Holds time of previous SuperCycle pulse
+	double		laser_prev;	///< Holds time of previous Laser status pulse
+	double		caen_time;	///< Time from the caen DAQ
+	double		caen_prev;	///< Holds previous time from the CAEN DAQ
+	std::vector<double> fpga_time; 			///< FPGA time on a given module of the array
+	std::vector<double> fpga_prev;				///< Previous FPGA time on a given module of the array
+	std::vector<double> asic_time;				///< ASIC time on a given module of the array
+	std::vector<double> asic_prev;				///< Previous ASIC time on a given module of the array
+	std::vector<double> pause_time;			///< The pause time on a given module of the array
+	std::vector<double> resume_time;			///< The resume time on a given module of the array
+	std::vector<double> asic_dead_time;		///< ASIC dead time for a given module of the array
+	std::vector<double> asic_time_start;		///< Holds the time of the first hit on each asic in the input time-sorted tree (index denotes asic module)
+	std::vector<double> asic_time_stop;		///< Holds the time of the last hit on each asic in the input time-sorted tree (index denotes asic module)
+	std::vector<double> caen_time_start;		///< Holds the time of the first hit on each caen in the input time-sorted tree (index denotes caen module)
+	std::vector<double> caen_time_stop;		///< Holds the time of the last hit on each asic in the input time-sorted tree (index denotes caen module)
 
 	// Data variables - generic
 	unsigned char		mymod;		///< module number
 	unsigned char		mych;		///< channel number
-	long double			mytime;		///< absolute timestamp
+	double			mytime;		///< absolute timestamp
 	float 				myenergy;	///< calibrated energy
 	int 				mywalk;		///< time walk correction
 	bool				mythres;	///< above threshold?
@@ -225,10 +226,10 @@ private:
 	// Array variables
 	std::vector<float>			pen_list;	///< list of p-side energies for ParticleFinder
 	std::vector<float>			nen_list;	///< list of n-side energies for ParticleFinder
-	std::vector<long double>	ptd_list;	///< list of p-side time differences for ParticleFinder without time walk correction
-	std::vector<long double>	ntd_list;	///< list of n-side time differences for ParticleFinder without time walk correction
-	std::vector<long double>	pwalk_list;	///< list of p-side time differences for ParticleFinder WITH time walk correction
-	std::vector<long double>	nwalk_list;	///< list of n-side time differences for ParticleFinder WITH time walk correction
+	std::vector<double>	ptd_list;	///< list of p-side time differences for ParticleFinder without time walk correction
+	std::vector<double>	ntd_list;	///< list of n-side time differences for ParticleFinder without time walk correction
+	std::vector<double>	pwalk_list;	///< list of p-side time differences for ParticleFinder WITH time walk correction
+	std::vector<double>	nwalk_list;	///< list of n-side time differences for ParticleFinder WITH time walk correction
 	std::vector<char>			pid_list;	///< list of p-side strip ids
 	std::vector<char>			nid_list;	///< list of n-side strip ids
 	std::vector<char>			pmod_list;	///< list of p-side modules numbers
@@ -240,29 +241,29 @@ private:
 
 	// Recoil variables
 	std::vector<float>			ren_list;	///< list of recoil energies for RecoilFinder
-	std::vector<long double>	rtd_list;	///< list of recoil time differences for RecoilFinder
+	std::vector<double>	rtd_list;	///< list of recoil time differences for RecoilFinder
 	std::vector<char>			rid_list;	///< list of recoil IDs/layers for RecoilFinder
 	std::vector<char>			rsec_list;	///< list of recoil sectors for RecoilFinder
 	
 	// MWPC variables
 	std::vector<unsigned short>	mwpctac_list;	///< TAC time from the MWPC
-	std::vector<long double>	mwpctd_list;	///< list of ELUM time differences for ELUMFinder
+	std::vector<double>	mwpctd_list;	///< list of ELUM time differences for ELUMFinder
 	std::vector<char>			mwpcaxis_list;	///< list of axis IDs for the MWPC
 	std::vector<char>			mwpcid_list;	///< list of TAC IDs for the MWPC
 
 	// ELUM variables
 	std::vector<float>			een_list;	///< list of ELUM energies for ELUMFinder
-	std::vector<long double>	etd_list;	///< list of ELUM time differences for ELUMFinder
+	std::vector<double>	etd_list;	///< list of ELUM time differences for ELUMFinder
 	std::vector<char>			esec_list;	///< list of ELUM sectors for ELUMFinder
 
 	// ZeroDegree variables
 	std::vector<float>			zen_list;	///< list of ZeroDegree energies for ELUMFinder
-	std::vector<long double>	ztd_list;	///< list of ZeroDegree time differences for ELUMFinder
+	std::vector<double>	ztd_list;	///< list of ZeroDegree time differences for ELUMFinder
 	std::vector<char>			zid_list;	///< list of ZeroDegree IDs/layers for ELUMFinder
 
 	// ScintArray variables
 	std::vector<float>			saen_list;	///< list of ScintArray energies for GammaFinder
-	std::vector<long double>	satd_list;	///< list of ScintArray time differences for GammaFinder
+	std::vector<double>	satd_list;	///< list of ScintArray time differences for GammaFinder
 	std::vector<char>			said_list;	///< list of ScintArray detectors ids for GammaFinder
 
 	// Counters
